// packages/shared/src/api/model/categoryResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Drones API
 * OpenAPI spec version: v1
 */

export interface CategoryResponse {
  id?: number;
  restaurantId?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
}