// packages/shared/src/api/model/droneResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Drones API
 * OpenAPI spec version: v1
 */

export interface DroneResponse {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  status?: string | null;
  batteryLevel?: number;
  maxPayloadKg?: number;
  currentLatitude?: number;
  currentLongitude?: number;
  /** @nullable */
  lastSeenAt?: string | null;
  createdAt?: string;
}