// packages/shared/src/api/model/productResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Drones API
 * OpenAPI spec version: v1
 */

export interface ProductResponse {
  id?: string;
  categoryId?: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  price?: number;
  weightKg?: number;
  isAvailable?: boolean;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
}