// packages/shared/src/api/model/restaurantResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Drones API
 * OpenAPI spec version: v1
 */

export interface RestaurantResponse {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  latitude?: number;
  longitude?: number;
  isOpen?: boolean;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  openTime?: string;
  closeTime?: string;
  averageRating?: number;
  reviewCount?: number;
  createdAt?: string;
}